plenty of good changes and bug fixes

+++ kernel/include/memory.hpp
@@ -2,10 +2,10 @@
 #include "dstd/cstdint.hpp"
 
 [[nodiscard]]
-void* operator new(uint32_t count);
+void* operator new(uint64_t count);
 
 [[nodiscard]]
-void* operator new[](uint32_t count);
+void* operator new[](uint64_t count);
 
 void operator delete(void* ptr);
 void operator delete[](void* ptr);
